Vim : Add command snippets to Vim and improve the display
This commit adds command snippets to Vim, which are displayed in an FZF window for easy selection. The snippets are defined in the `command_snippets` file. The display of the snippets has been improved with better formatting, including adding a separator between the subject and command, and adding spaces to make the display more readable. The commit also includes several new snippets for reloading opened files, swapping left and right panels, and displaying the next and previous quickfix panels.

modified:   vim/.vimrc
modified:   vim/functions/command_snippets