Vim : Add functions and commands to navigate Quickfix list in Vim
This commit adds two functions and commands to navigate the Quickfix list in Vim.
The `NextQuickfix` function and command navigate to the next item in the list, while the `PreviousQuickfix` function and command navigate to the previous item.
The new commands are mapped to `<leader>'` and `<leader>;` respectively.

modified:   vim/.vimrc