refactor(vim): remove SearchOrOpenInCloudNote command
This commit removes the SearchOrOpenInCloudNote command and its
associated function from the .vimrc and the vim/functions/normal
file. The decision to remove this feature is based on its limited
usage and the desire to simplify the Vim configuration. This change
aims to enhance the maintainability and performance of the Vim
environment.

modified:   vim/.vimrc
modified:   vim/functions/normal