feat: update CopilotChatCommitStaged handling
This change modifies the way CopilotChat interacts with git commit
messages. Instead of using the FileType event for "gitcommit", it now
triggers on BufRead for the COMMIT_EDITMSG file. This workaround is
implemented due to an issue where the prompt was executed multiple
times, likely a bug in Neovim 10.

modified:   vim/lua_scripts/init/copilot_chat.lua